make within main now runs with no compilation errors and few warnings (using vanilla compiler flags).
Undefined references on linking.

git-svn-id: http://projects.cs.kent.ac.uk/projects/cxxr/svn/trunk@9 549f2c98-2231-0410-b618-a09cd5e121c0

+++ src/main/colors.cpp
@@ -270,9 +270,9 @@ SEXP attribute_hidden do_hcl(SEXP call, SEXP op, SEXP args, SEXP env)
         if (L < 0 || L > WHITE_Y || C < 0 || A < 0 || A > 1)
             errorcall(call, _("invalid hcl color"));
         hcl2rgb(H, C, L, &r, &g, &b);
-        ir = 255 * r + .5;
-        ig = 255 * g + .5;
-        ib = 255 * b + .5;
+        ir = int(255 * r + .5);
+        ig = int(255 * g + .5);
+        ib = int(255 * b + .5);
         if (FixupColor(&ir, &ig, &ib) && !fixup)
             SET_STRING_ELT(ans, i, NA_STRING);
         else
@@ -365,7 +365,7 @@ SEXP attribute_hidden do_gray(SEXP call, SEXP op, SEXP args, SEXP env)
 	level = REAL(lev)[i];
 	if (ISNAN(level) || level < 0 || level > 1)
 	    errorcall(call, _("invalid gray level, must be in [0,1]."));
-	ilevel = 255 * level + 0.5;
+	ilevel = int(255 * level + 0.5);
 	SET_STRING_ELT(ans, i, mkChar(RGB2rgb(ilevel, ilevel, ilevel)));
     }
     UNPROTECT(2);
